+++ vscode/src/chat/chat-view/SimpleChatPanelProvider.ts
@@ -24,10 +24,13 @@ import { Result } from '@sourcegraph/cody-shared/src/local-context'
 import { MAX_BYTES_PER_FILE, NUM_CODE_RESULTS, NUM_TEXT_RESULTS } from '@sourcegraph/cody-shared/src/prompt/constants'
 import { truncateTextNearestLine } from '@sourcegraph/cody-shared/src/prompt/truncation'
 import { Message } from '@sourcegraph/cody-shared/src/sourcegraph-api'
+import { isDotCom } from '@sourcegraph/cody-shared/src/sourcegraph-api/environments'
+import { ContextWindowLimitError, isRateLimitError } from '@sourcegraph/cody-shared/src/sourcegraph-api/errors'
 import { isError } from '@sourcegraph/cody-shared/src/utils'
 
 import { View } from '../../../webviews/NavBar'
 import { getFullConfig } from '../../configuration'
+import { executeEdit } from '../../edit/execute'
 import { getFileContextFiles, getOpenTabsContextFile, getSymbolContextFiles } from '../../editor/utils/editor-context'
 import { VSCodeEditor } from '../../editor/vscode-editor'
 import { ContextStatusAggregator } from '../../local-context/enhanced-context-status'
@@ -48,20 +51,15 @@ import {
 import { openExternalLinks, openFilePath, openLocalFileWithRange } from '../../services/utils/workspace-action'
 import { CachedRemoteEmbeddingsClient } from '../CachedRemoteEmbeddingsClient'
 import { MessageErrorType } from '../MessageProvider'
-import { ConfigurationSubsetForWebview, ExtensionMessage, LocalEnv, WebviewMessage } from '../protocol'
+import { AuthStatus, ConfigurationSubsetForWebview, ExtensionMessage, LocalEnv, WebviewMessage } from '../protocol'
 import { countGeneratedCode } from '../utils'
 
-import {
-    CodebaseStatusProvider,
-    embeddingsUrlScheme,
-    getChatPanelTitle,
-    relativeFileUrl,
-    stripContextWrapper,
-} from './chat-helpers'
+import { embeddingsUrlScheme, getChatPanelTitle, relativeFileUrl, stripContextWrapper } from './chat-helpers'
 import { ChatHistoryManager } from './ChatHistoryManager'
 import { addWebviewViewHTML, CodyChatPanelViewType } from './ChatManager'
 import { ChatViewProviderWebview } from './ChatPanelProvider'
 import { Config, IChatPanelProvider } from './ChatPanelsManager'
+import { CodebaseStatusProvider } from './CodebaseStatusProvider'
 import { InitDoer } from './InitDoer'
 import { DefaultPrompter, IContextProvider, IPrompter } from './prompt'
 import { ContextItem, MessageWithContext, SimpleChatModel, toViewMessage } from './SimpleChatModel'
@@ -308,6 +306,9 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
                 logDebug('SimpleChatPanelProvider:onDidReceiveMessage', 'initialized')
                 await this.onInitialized()
                 break
+            case 'reset':
+                await this.clearAndRestartSession()
+                break
             case 'submit': {
                 const requestID = uuid.v4()
                 await this.handleHumanMessageSubmitted(
@@ -375,11 +376,19 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
             case 'embeddings/index':
                 void this.localEmbeddings?.index()
                 break
+            case 'symf/index': {
+                void this.codebaseStatusProvider.currentCodebase().then((codebase): void => {
+                    if (codebase) {
+                        void this.symf?.ensureIndex(codebase.local, { hard: true })
+                    }
+                })
+                break
+            }
             case 'show-page':
                 await vscode.commands.executeCommand('cody.show-page', message.page)
                 break
             default:
-                this.postError(new Error('Invalid request type from Webview Panel'))
+                this.postError(new Error(`Invalid request type from Webview Panel: ${message.command}`))
         }
     }
 
@@ -444,7 +453,10 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
         if (this.chatModel.isEmpty()) {
             return
         }
+
+        this.cancelInProgressCompletion()
         await this.saveSession()
+
         this.chatModel = new SimpleChatModel(this.chatModel.modelID)
         this.sessionID = this.chatModel.sessionID
         this.postViewTranscript()
@@ -503,10 +515,24 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
             const args = { requestID }
             telemetryService.log('CodyVSCodeExtension:chatPredictions:used', args, { hasV2Event: true })
         }
+
         // If this is a slash command, run it with custom prompt recipe instead
         if (text.startsWith('/')) {
             if (text.match(/^\/r(eset)?$/)) {
                 await this.clearAndRestartSession()
+                return
+            }
+            if (text.match(/^\/edit(\s)?/)) {
+                await executeEdit({ instruction: text.replace(/^\/(edit)/, '').trim() }, 'chat')
+                return
+            }
+            if (text.match(/^\/doc(\s)?/)) {
+                await vscode.commands.executeCommand('cody.command.document-code')
+                return
+            }
+            if (text === '/commands-settings') {
+                // User has clicked the settings button for commands
+                await vscode.commands.executeCommand('cody.settings.commands')
                 return
             }
             return this.executeRecipe('custom-prompt', text.trim(), 'chat', userContextFiles, addEnhancedContext)
@@ -571,7 +597,10 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
         sendTelemetry?: (contextSummary: {}) => void
     ): Promise<void> {
         try {
-            const contextWindowBytes = 28000 // 7000 tokens * 4 bytes per token
+            const contextWindowBytes = getContextWindowForModel(
+                this.authProvider.getAuthStatus(),
+                this.chatModel.modelID
+            )
 
             const userContextItems = await contextFilesToContextItems(this.editor, userContextFiles || [], true)
             const contextProvider = new ContextProvider(
@@ -582,7 +611,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
                 this.config.experimentalSymfContext ? this.symf : null,
                 this.codebaseStatusProvider
             )
-            const { prompt, warnings, newContextUsed } = await this.prompter.makePrompt(
+            const { prompt, contextLimitWarnings, newContextUsed } = await this.prompter.makePrompt(
                 this.chatModel,
                 contextProvider,
                 addEnhancedContext,
@@ -591,10 +620,11 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
 
             this.chatModel.setNewContextUsed(newContextUsed)
 
-            if (warnings.length > 0) {
-                const warningMsg =
-                    'Warning: ' + warnings.map(w => (w.trim().endsWith('.') ? w.trim() : w.trim() + '.')).join(' ')
-                this.postError(new Error(warningMsg))
+            if (contextLimitWarnings.length > 0) {
+                const warningMsg = contextLimitWarnings
+                    .map(w => (w.trim().endsWith('.') ? w.trim() : w.trim() + '.'))
+                    .join(' ')
+                this.postError(new ContextWindowLimitError(warningMsg), 'transcript')
             }
 
             if (sendTelemetry) {
@@ -641,7 +671,11 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
                 },
             })
         } catch (error) {
-            this.postError(new Error(`Error generating assistant response: ${error}`))
+            if (isRateLimitError(error)) {
+                this.postError(error, 'transcript')
+            } else {
+                this.postError(isError(error) ? error : new Error(`Error generating assistant response: ${error}`))
+            }
         }
     }
 
@@ -763,7 +797,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
             return
         }
         // Update webview panel title to match the last message
-        const text = this.chatModel.getLastHumanMessages()?.displayText
+        const text = this.chatModel.getLastHumanMessage()?.displayText
         if (this.webviewPanel && text) {
             this.webviewPanel.title = getChatPanelTitle(text)
         }
@@ -776,6 +810,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
         // Add error to transcript
         if (type === 'transcript') {
             this.chatModel.addErrorAsBotMessage(error)
+            this.postViewTranscript()
             void this.postMessage({ type: 'transcript-errors', isTranscriptError: true })
             return
         }
@@ -923,8 +958,15 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
             const allCommands = await this.editor.controllers.command?.getAllCommands(true)
             // HACK: filter out commands that make inline changes and /ask (synonymous with a generic question)
             const prompts =
-                allCommands?.filter(([id, { mode = '' }]) => {
-                    return !['/edit', '/doc', '/ask'].includes(id) && !['edit', 'insert'].includes(mode)
+                allCommands?.filter(([id, { mode }]) => {
+                    /** The /ask command is only useful outside of chat */
+                    const isRedundantCommand = id === '/ask'
+                    /**
+                     * Hack: Custom edit commands are currently broken in this chat.
+                     * We filter our anything that has this mode, apart from our own internal doc command - which we override ourselves
+                     */
+                    const isCustomEdit = (mode === 'edit' || mode === 'insert') && id !== '/doc'
+                    return !isRedundantCommand && !isCustomEdit
                 }) || []
 
             void this.postMessage({
@@ -1008,16 +1050,21 @@ class ContextProvider implements IContextProvider {
         let remoteEmbeddingsError
         searchContext.push(...(await this.getReadmeContext()))
         logDebug('SimpleChatPanelProvider', 'getEnhancedContext > embeddings (start)')
+        let hasEmbeddingsContext = false
         const localEmbeddingsResults = this.searchEmbeddingsLocal(text)
         const remoteEmbeddingsResults = this.searchEmbeddingsRemote(text)
         try {
-            searchContext.push(...(await localEmbeddingsResults))
+            const r = await localEmbeddingsResults
+            hasEmbeddingsContext = hasEmbeddingsContext || r.length > 0
+            searchContext.push(...r)
         } catch (error) {
             logDebug('SimpleChatPanelProvider', 'getEnhancedContext > local embeddings', error)
             localEmbeddingsError = error
         }
         try {
-            searchContext.push(...(await remoteEmbeddingsResults))
+            const r = await remoteEmbeddingsResults
+            hasEmbeddingsContext = hasEmbeddingsContext || r.length > 0
+            searchContext.push(...r)
         } catch (error) {
             logDebug('SimpleChatPanelProvider', 'getEnhancedContext > remote embeddings', error)
             remoteEmbeddingsError = error
@@ -1031,9 +1078,14 @@ class ContextProvider implements IContextProvider {
             )
         }
 
-        if (searchContext.length === 0 && this.symf) {
-            // Fallback to symf if embeddings provided no results
-            searchContext.push(...(await this.searchSymf(text)))
+        if (!hasEmbeddingsContext && this.symf) {
+            try {
+                // Fallback to symf if embeddings provided no results
+                searchContext.push(...(await this.searchSymf(text)))
+            } catch (error) {
+                // TODO(beyang): handle this error better
+                logDebug('SimpleChatPanelProvider.getEnhancedContext', 'searchSymf error', error)
+            }
         }
 
         const priorityContext: ContextItem[] = []
@@ -1064,12 +1116,18 @@ class ContextProvider implements IContextProvider {
     /**
      * Uses symf to conduct a local search within the current workspace folder
      */
-    private async searchSymf(userText: string): Promise<ContextItem[]> {
+    private async searchSymf(userText: string, blockOnIndex = false): Promise<ContextItem[]> {
         if (!this.symf) {
             return []
         }
         const workspaceRoot = this.editor.getWorkspaceRootUri()?.fsPath
         if (!workspaceRoot) {
+            return []
+        }
+
+        const indexExists = await this.symf.getIndexStatus(workspaceRoot)
+        if (indexExists !== 'ready' && !blockOnIndex) {
+            void this.symf.ensureIndex(workspaceRoot, { hard: false })
             return []
         }
 
@@ -1408,4 +1466,37 @@ function getErrorMessage(error: unknown): string {
         return error.message
     }
     return String(error)
+}
+
+function getContextWindowForModel(authStatus: AuthStatus, modelID: string): number {
+    // In enterprise mode, we let the sg instance dictate the token limits and allow users to
+    // overwrite it locally (for debugging purposes).
+    //
+    // This is similiar to the behavior we had before introducing the new chat and allows BYOK
+    // customers to set a model of their choice without us having to map it to a known model on
+    // the client.
+    if (authStatus.endpoint && !isDotCom(authStatus.endpoint)) {
+        const codyConfig = vscode.workspace.getConfiguration('cody')
+        const tokenLimit = codyConfig.get<number>('provider.limit.prompt')
+        if (tokenLimit) {
+            return tokenLimit * 4 // bytes per token
+        }
+
+        if (authStatus.configOverwrites?.chatModelMaxTokens) {
+            return authStatus.configOverwrites.chatModelMaxTokens * 4 // butes per token
+        }
+
+        return 28000 // 7000 tokens * 4 bytes per token
+    }
+
+    if (modelID.includes('openai/gpt-4-1106-preview')) {
+        return 28000 // 7000 tokens * 4 bytes per token
+    }
+    if (modelID.endsWith('openai/gpt-3.5-turbo')) {
+        return 10000 // 4,096 tokens * < 4 bytes per token
+    }
+    if (modelID.includes('mixtral-8x7b-instruct') && modelID.includes('fireworks')) {
+        return 28000 // 7000 tokens * 4 bytes per token
+    }
+    return 28000 // assume default to Claude-2-like model
 }