+++ vscode/webviews/App.tsx
@@ -66,6 +66,9 @@ export const App: React.FunctionComponent<{ vscodeAPI: VSCodeWrapper }> = ({ vsc
     const onConsentToEmbeddings = useCallback((): void => {
         vscodeAPI.postMessage({ command: 'embeddings/index' })
     }, [vscodeAPI])
+    const onShouldBuildSymfIndex = useCallback((): void => {
+        vscodeAPI.postMessage({ command: 'symf/index' })
+    }, [vscodeAPI])
 
     useEffect(
         () =>
@@ -240,6 +243,7 @@ export const App: React.FunctionComponent<{ vscodeAPI: VSCodeWrapper }> = ({ vsc
                         <EnhancedContextEventHandlers.Provider
                             value={{
                                 onConsentToEmbeddings,
+                                onShouldBuildSymfIndex,
                                 onEnabledChange: (enabled): void => {
                                     if (enabled !== enhancedContextEnabled) {
                                         setEnhancedContextEnabled(enabled)
@@ -350,7 +354,7 @@ You can start a new chat at any time with <span class="keyboard-shortcut"><span>
         os === 'darwin' ? '⌥' : 'Alt'
     }</span><span>/</span></span> or using the <span class="cody-icons">H</span> button.
 
-For more tips and tricks, see the [Getting Started Guide](command:cody.welcome) and [docs](https://docs.sourcegraph.com/cody).
+For more tips and tricks, see the [Getting Started Guide](command:cody.welcome) and [docs](https://sourcegraph.com/docs/cody).
 `
     return welcomeMessageMarkdown
 }